# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TaleMotion Backend Foundation API
  version: 0.1.0
  description: PostgreSQL-backed project, scene, and job foundation.
servers:
  - url: http://localhost:8000/api/v1
security:
  - cookieAuth: []
paths:
...
  /chapters/{chapter_id}/scenes:
    post:
      tags: [Scenes]
      summary: Add or insert a scene
      parameters:
        - {$ref: "#/components/parameters/ChapterId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateSceneRequest"}
      responses:
        "201":
          description: Scene
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SceneResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /chapters/{chapter_id}/scenes/reorder:
    post:
      tags: [Scenes]
      summary: Reorder every scene in a chapter
      parameters:
        - {$ref: "#/components/parameters/ChapterId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ReorderScenesRequest"}
      responses:
        "200":
          description: Updated chapter
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ChapterResponse"}
        "409": {$ref: "#/components/responses/Error"}
  /scenes/{scene_id}:
    parameters:
      - {$ref: "#/components/parameters/SceneId"}
    get:
      tags: [Scenes]
      summary: Get a scene
      responses:
        "200":
          description: Scene
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SceneResponse"}
        "404": {$ref: "#/components/responses/Error"}
    patch:
      tags: [Scenes]
      summary: Update editable scene fields
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateSceneRequest"}
      responses:
        "200":
          description: Updated scene
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SceneResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
    delete:
      tags: [Scenes]
      summary: Delete a scene and normalize positions
      responses:
        "204": {description: Scene deleted}
        "404": {$ref: "#/components/responses/Error"}
  /scenes/{scene_id}/duplicate:
    post:
      tags: [Scenes]
      summary: Duplicate a scene after its source
      parameters:
        - {$ref: "#/components/parameters/SceneId"}
      responses:
        "201":
          description: Draft duplicate
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SceneResponse"}
        "404": {$ref: "#/components/responses/Error"}
  /scenes/{scene_id}/generations:
    post:
      tags: [Scenes]
      summary: Queue real media generation for a persisted scene
      parameters:
        - {$ref: "#/components/parameters/SceneId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateSceneGenerationRequest"}
      responses:
        "202":
          description: Persisted queued generation job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /scenes/{scene_id}/regenerations:
    post:
      tags: [Scenes]
      summary: Queue a new scene media version
      parameters:
        - {$ref: "#/components/parameters/SceneId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateSceneRegenerationRequest"}
      responses:
        "202":
          description: Persisted queued regeneration job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: talemotion_session
  parameters:
...
    ChapterId: {name: chapter_id, in: path, required: true, schema: {type: string}}
    SceneId: {name: scene_id, in: path, required: true, schema: {type: string}}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 1, maxLength: 128}
...
  responses:
    Error:
      description: Consistent error envelope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
  schemas:
...
    ChapterStatus:
      type: string
      enum: [draft, ready, generating, completed, failed]
    SceneStatus:
      type: string
      enum: [draft, ready, queued, generating, completed, failed]
    JobType:
      type: string
      enum: [storyboard, scene_generation, scene_regeneration, project_generation, render]
    JobStatus:
      type: string
      enum: [queued, running, completed, failed, cancel_requested, cancelled]
...
    ChapterResponse:
      type: object
      required: [id, project_id, title, position, status, scenes, created_at, updated_at]
      properties:
        id: {type: string}
        project_id: {type: string}
        title: {type: string}
        summary: {type: [string, "null"]}
        position: {type: integer, minimum: 1}
        target_duration_seconds: {type: [integer, "null"]}
        status: {$ref: "#/components/schemas/ChapterStatus"}
        scenes: {type: array, items: {$ref: "#/components/schemas/SceneResponse"}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateSceneRequest:
      type: object
      additionalProperties: false
      required: [title, narration, visual_prompt, duration_seconds]
      properties:
        title: {type: string, minLength: 1}
        narration: {type: string}
        visual_prompt: {type: string}
        duration_seconds: {type: integer, minimum: 1, maximum: 60}
        position: {type: integer, minimum: 1}
    CreateSceneGenerationRequest:
      type: object
      additionalProperties: false
      properties:
        duration_seconds: {type: integer, minimum: 1, maximum: 60, default: 5}
        generate_video: {type: boolean, default: true}
    CreateSceneRegenerationRequest:
      type: object
      additionalProperties: false
      required: [additional_instruction]
      properties:
        additional_instruction: {type: string, minLength: 1, maxLength: 2000}
        duration_seconds: {type: integer, minimum: 1, maximum: 60, default: 5}
        generate_video: {type: boolean, default: true}
...
    UpdateSceneRequest:
      type: object
      additionalProperties: false
      properties:
        title: {type: string, minLength: 1}
        narration: {type: string}
        visual_prompt: {type: string}
        duration_seconds: {type: integer, minimum: 1, maximum: 60}
    ReorderScenesRequest:
      type: object
      required: [scene_ids]
      properties:
        scene_ids: {type: array, items: {type: string}}
    SceneResponse:
      allOf:
        - {$ref: "#/components/schemas/CreateSceneRequest"}
        - type: object
          required: [id, chapter_id, position, status, active_asset_version, created_at, updated_at]
          properties:
            id: {type: string}
            chapter_id: {type: string}
            status: {$ref: "#/components/schemas/SceneStatus"}
            active_asset_id: {type: [string, "null"]}
            active_asset_version: {type: integer}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
    GenerationJobResponse:
      type: object
      required: [id, project_id, type, status, progress, input_payload, retry_count, max_retries, created_at, updated_at, children]
      properties:
        id: {type: string}
        project_id: {type: string}
        scene_id: {type: [string, "null"]}
        parent_job_id: {type: [string, "null"]}
        type: {$ref: "#/components/schemas/JobType"}
        status: {$ref: "#/components/schemas/JobStatus"}
        progress: {type: integer, minimum: 0, maximum: 100}
        current_stage: {type: [string, "null"]}
        input_payload: {type: object}
        result_payload: {type: [object, "null"]}
        error_code: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
        retry_count: {type: integer}
        max_retries: {type: integer}
        cancel_requested_at: {type: [string, "null"], format: date-time}
        created_at: {type: string, format: date-time}
        started_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        updated_at: {type: string, format: date-time}
        children:
          type: array
          items: {$ref: "#/components/schemas/GenerationJobChildResponse"}
...
    GenerationJobChildResponse:
      type: object
      required: [id, scene_id, status, progress, result_asset_id, error_code, error_message]
      properties:
        id: {type: string}
        scene_id: {type: [string, "null"]}
        status: {$ref: "#/components/schemas/JobStatus"}
        progress: {type: integer, minimum: 0, maximum: 100}
        result_asset_id: {type: [string, "null"]}
        error_code: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            details: {type: object}
            request_id: {type: string}